+++ src/main/java/cz/cuni/mff/kotal/frontend/simulation/AbstractGraph.java
@@ -14,12 +14,12 @@ import java.util.Set;
 public class AbstractGraph extends SimulationGraph {
 	private final IntersectionMenuTab0.Parameters.Models model;
 
-	public AbstractGraph(IntersectionMenuTab0.Parameters.Models model, long granularity, long entries, long exits, boolean oriented, Set<GraphicalVertex> vertices, Map<Integer, List<Vertex>> entryExitVertices, Set<Edge> edges) {
+	public AbstractGraph(IntersectionMenuTab0.Parameters.Models model, int granularity, int entries, int exits, boolean oriented, Set<GraphicalVertex> vertices, Map<Integer, List<Vertex>> entryExitVertices, Set<Edge> edges) {
 		super(granularity, entries, exits, oriented, vertices, entryExitVertices, edges);
 		this.model = model;
 	}
 
-	public AbstractGraph(IntersectionMenuTab0.Parameters.Models model, long granularity, long entries, long exits, boolean oriented) {
+	public AbstractGraph(IntersectionMenuTab0.Parameters.Models model, int granularity, int entries, int exits, boolean oriented) {
 		this(model, granularity, entries, exits, oriented, null, null, null);
 	}
