+++ src/main/java/cz/cuni/mff/kotal/frontend/simulation/AgentPane.java
@@ -27,7 +27,7 @@ import static cz.cuni.mff.kotal.helpers.MyGenerator.generateRandomInt;
 public class AgentPane extends StackPane {
 
 	public final Color color = Color.rgb(generateRandomInt(255), generateRandomInt(255), generateRandomInt(255));
-	private final Map<Long, Vertex> simulationVertices;  // FIXME remove
+	private final Map<Integer, Vertex> simulationVertices;  // FIXME remove
 	private final Rectangle rectangle;
 	private final Rotate rotation = new Rotate();
 	private final Agent agent;
@@ -54,7 +54,7 @@ public class AgentPane extends StackPane {
 	 * @param simulationVertices Intersection vertices the agent is travelling on
 	 */
 	@Deprecated
-	public AgentPane(long startTime, double firstStep, Agent agent, double period, Map<Long, Vertex> simulationVertices, double cellSize) {
+	public AgentPane(long startTime, double firstStep, Agent agent, double period, Map<Integer, Vertex> simulationVertices, double cellSize) {
 		this.agent = agent;
 		this.distanceTraveled = 0;
 		this.startTime = startTime;
@@ -69,7 +69,7 @@ public class AgentPane extends StackPane {
 		}
 	}
 
-	public AgentPane(double firstStep, Agent agent, Map<Long, Vertex> simulationVertices, double cellSize) {
+	public AgentPane(double firstStep, Agent agent, Map<Integer, Vertex> simulationVertices, double cellSize) {
 		this.agent = agent;
 		this.simulationVertices = simulationVertices;
 
@@ -160,7 +160,7 @@ public class AgentPane extends StackPane {
 	 * @return
 	 */
 	private boolean computeNewAngle(double time) {
-		Pair<Long, Long> previousNext = agent.getPreviousNextVertexIDs(time);
+		Pair<Integer, Integer> previousNext = agent.getPreviousNextVertexIDs(time);
 		GraphicalVertex start = (GraphicalVertex) simulationVertices.get(previousNext.getKey());
 		GraphicalVertex end = (GraphicalVertex) simulationVertices.get(previousNext.getValue());
